move CustomBasicAuth override to import
